LayoutTests/http/tests/site-isolation/accessibility/client/simple-iframe.html

<html><!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<head>
<script src="/js-test-resources/js-test.js"></script>
<script src="/js-test-resources/accessibility-helper.js"></script>
</head>
<body>

<button>Before</button>

<iframe id="iframe" src="http://localhost:8000/site-isolation/resources/iframe.html"></iframe>

<button>After</button>

<script>
var output = "Dump accessibility tree with isolated iframe\n";

if (window.accessibilityController) {
    window.jsTestIsAsync = true;
    accessibilityController.setClientAccessibilityMode(true);

    setTimeout(async function() {
        // Wait for both a button and heading to appear in the accessibility tree.
        await waitForElements([
            (element) => element.role.includes("Button"),
            (element) => element.role.includes("Heading"),
        ]);

        const allowedRoles = ["AXWebArea", "AXButton", "AXHeading"];
        function dumpFilteredAccessibilityTree(element, indent) {
            const role = element.role;
            const matched = allowedRoles.some(r => role.endsWith(r));
            if (matched) {
                for (let i = 0; i < indent; i++)
                    output += "    ";
                output += role + " " + element.title + "\n";
            }
            const childIndent = matched ? indent + 1 : indent;
            for (let i = 0; i < element.childrenCount; i++)
                dumpFilteredAccessibilityTree(element.childAtIndex(i), childIndent);
        }
        dumpFilteredAccessibilityTree(accessibilityController.rootElement, 0);
        debug(output);
        finishJSTest();
    }, 0);
}
</script>
</body>
</html>
